refactor: Make fn fetch_events
